==> Old_Rmd_files/05_C57BL6J_2DG_WGCNA_summary_SM_SI_spleen_96hrs.Rmd <==
---
title: "Summary Skeletal Muscle, Small Intestine, and Spleen 96hrs"
author: Ann Wells
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
 html_document:
   keep_md: TRUE
   code_folding: hide
   collapsed: no
   toc: yes
   toc_float: yes
   self_contained: no
   
---


# Summary

```{r setup, echo=FALSE, warning=FALSE, message=FALSE}
options(stringsAsFactors = F)
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
knitr::opts_knit$set(progress=FALSE)

```

```{r load packages}
needed.packages <- c("tidyverse", "here", "functional", "gplots", "dplyr", "GeneOverlap", "R.utils", "reshape2","magrittr","data.table", "RColorBrewer","preprocessCore", "ARTool","emmeans", "phia", "gProfileR","rlist", "pheatmap")
for(i in 1:length(needed.packages)){library(needed.packages[i], character.only = TRUE)}

source(here("source_files","WGCNA_source.R"))
```

There were 47 samples (4 per time and treatment minus skeletal muscle 96 hr control). Mice were fed 2DG (6g/L) or nothing in their water for 96hrs or 2 weeks. WGCNA analysis was performed to determine gene clustering. A soft-thresholding power was chosen so that the scale-free topology correlation attempts to hit 0.9. The power chosen for skeletal muscle, small intestine, and spleen for 96 hrs was 1, which allowed the correlations to hover around 0.98. This power clustered the genes into 5 modules originally, however, modules that were highly correlated (0.75 or higher) were further merged together to create 4 modules.

```{r}
modules<-read.csv(here("Data","SM.SI.spleen.96hrs","log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs.WGCNA.module.membership.csv"), header=T)

module_barplot(modules)
```

Pathways were assessed for each module. Out of 4 modules, only 3 modules had pathway hits of any kind. There were many pathways that had the highest frequency across modules, including metabolic pathways, focal adhesion, axon guidance, and cell cycle. The blue, brown, and turquoise modules all contained these pathways.

## Blue

182 pathways were identified for the **blue** module. The pathways identified were involved in cell cycle, external cellular response (hypoxia, stress), Class 1 MHC, glycogen metabolism, cancer pathways, and immune pathways.

## Brown

69 pathways were identified for the **brown** module. The pathways identified were involved in amin acid metabolism, cholesterol biosynthesis, glutathione, galactose, fructose, and glycerolipid metabolism, steroid and triglyceride biosynthesis.

## Turquoise

140 pathways were identified for the **turquoise** module. The pathways identified were involved in Alzheimer's disease, chromatin, DNA and RNA pathways, NAFLD, oxidative phosphorylation, and toll like receptor pathways.

```{r}
tdata.FPKM.sample.info <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_phenotypes.RData"))

tdata.FPKM <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_numeric.RData"))

log.tdata.FPKM <- log(tdata.FPKM + 1)
log.tdata.FPKM <- as.data.frame(log.tdata.FPKM)

log.tdata.FPKM.sample.info <- cbind(log.tdata.FPKM, tdata.FPKM.sample.info[,27238:27240])

log.tdata.FPKM.sample.info <- log.tdata.FPKM.sample.info %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.subset <- log.tdata.FPKM[,colMeans(log.tdata.FPKM != 0) > 0.5] 

log.tdata.FPKM.subset <- log.tdata.FPKM.subset %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs <- log.tdata.FPKM.sample.info %>% rownames_to_column() %>% filter(Tissue %in% c("Skeletal Muscle", "Spleen", "Small Intestine")) %>% filter(Time =="96 hrs") %>% column_to_rownames()

log.tdata.FPKM.subset <- subset(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs, select = -c(Time,Treatment,Tissue))
  
WGCNA.pathway <-readRDS(here("Data","SM.SI.spleen.96hrs","Chang_B6_96hr_4wk_gprofiler_pathway_annotation_list_SM_SI_spleen_96hrs_WGCNA.RData"))

Matched<-readRDS(here("Data","SM.SI.spleen.96hrs","Annotated_genes_in_SM_SI_spleen_96hrs_WGCNA_Chang_B6_96hr_4wk.RData"))


module.names <- Matched$X..Module.
name <- strsplit(module.names,"_")
samples <-c()
for(i in 1:length(name)){
samples[[i]] <- name[[i]][2]
}  
name <- strsplit(samples,"\"")
name <- unlist(name)

Treatment <- unclass(as.factor(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs[,27238]))
Tissue <- unclass(as.factor(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs[,27239]))
Treat.Tissue <- paste0(Treatment,Tissue)

phenotype <- data.frame(cbind(Tissue, Treatment,Treat.Tissue))

nSamples <- nrow(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs)

MEs0 <- read.csv(here("Data","SM.SI.spleen.96hrs","log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs.WGCNA.module.eigens.csv"),header = T, row.names = 1)
name <- strsplit(names(MEs0),"_")
samples <-c()
for(i in 1:length(name)){
samples[[i]] <- name[[i]][2]
}  
name <- strsplit(samples,"\"")
name <- unlist(name)
colnames(MEs0) <-name
MEs <- orderMEs(MEs0)
moduleTraitCor <- cor(MEs, phenotype, use = "p");
moduleTraitPvalue <- corPvalueStudent(moduleTraitCor, nSamples)
```

## Relationship between Modules and Traits
```{r, results='asis'}
#sizeGrWindow(10,6)
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
                        signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)

# Display the correlation values within a heatmap plot
heat <- pheatmap(moduleTraitCor, main = paste("Module-trait relationships"), display_numbers = textMatrix,color=colorRampPalette(brewer.pal(n = 11, name = "Paired"))(10), cluster_rows = F, cluster_cols = F, fontsize_number = 4, angle_col = 45, number_color = "black", border_color = "white")
heat
```

## Tissue

The top 10 genes that were significant for tissue within the **brown** module and had the highest correlation with brown were:

1. Mpzl2
2. Muc13
3. Olfm4
4. 2610528J11Rik
5. Cdkn2b
6. Aoc1
7. Lgals2
8. Krt20
9. Adtrp
10. Sema4g

When assessing the overall sample contribution through their summary eigengenes it appears that for **brown** all intestine had a positive effect and spleen and skeletal muscle had a negative effect. Spleen and muscle had mostly under-expressed genes while intestine had mostly over-expressed genes.

## Treatment

The top 10 genes that were significant for treatment within the **brown** module and had the highest correlation with brown were:

1. Elovl7
2. Grb7
3. Mpzl2
4. Muc13
5. Olfm4
6. 2610528J11Rik
7. Spire2
8. Espn
9. Trim2
10. Cdkn2b

When assessing the overall sample contribution through their summary eigengenes it appears that for **brown** all intestine had a positive effect and spleen and skeletal muscle had a negative effect, all independent of treatment. Spleen and muscle had mostly under-expressed genes while intestine had mostly over-expressed genes.

The top 10 genes that were significant for treatment within the **yellow** module and had the highest correlation with yellow were:

1. Cst6
2. Cep170
3. Tspan8
4. Rbp7
5. Rev1
6. Rcsd1
7. Mtch2
8. Tor1aip1
9. Sord
10. Pitpnm2

When assessing the overall sample contribution through their summary eigengenes it appears that for **yellow** all spleen had a positive effect, intestine had a negative effect, all independent of treatment, and skeletal muscle mostly had a negative effect with two 2DG samples having a positive effect. Intestine and muscle had mostly under-expressed genes while spleen had mostly over-expressed genes.

The top 10 genes that were significant for treatment within the **turquoise** module and had the highest correlation with turquoise were:

1. Bbx
2. Malt1
3. Gabbr1
4. Chic2
5. Pogz
6. Fam49b
7. Terf2
8. Zfp263
9. Sft2d3
10. Fchsd2

When assessing the overall sample contribution through their summary eigengenes it appears that for **turquoise** all spleen had a positive effect and intestine and skeletal muscle had a negative effect, all independent of treatment. Intestine and muscle had mostly under-expressed genes while spleen had mostly over-expressed genes.

The top 10 genes that were significant for treatment within the **blue** module and had the highest correlation with blue were:

1. Smco1
2. Fbxl19
3. Itgb1bp2
4. Pacsin3
5. Rtn2
6. Stac3
7. Arhgap27
8. 2310015B20Rik
9. Sec23b
10. Hspb3

When assessing the overall sample contribution through their summary eigengenes it appears that for **blue** all intestine and spleen had a positive effect while skeletal muscle had a negative effect, all independent of treatment. Muscle had mostly under-expressed genes while intestine and spleen had mostly over-expressed genes.

## Treatment by tissue

The top 10 genes that were significant for treatment by tissue within the **brown** module and had the highest correlation with brown were:

1. Grb7
2. Mpzl2
3. Muc13
4. Olfm4
5. 2610528J11Rik
6. Spire2
7. Espn
8. Cdkn2b
9. Lgals2
10. Krt20

When assessing the overall sample contribution through their summary eigengenes it appears that for **brown** all intestine had a positive effect and spleen and skeletal muscle had a negative effect, all independent of treatment but not tissue. Spleen and muscle had mostly under-expressed genes while intestine had mostly over-expressed genes.

The top 10 genes that were significant for treatment by tissue within the **yellow** module and had the highest correlation with yellow were:

1. Ado
2. Fmnl2
3. Dhx29
4. Klhl24
5. Tulp3
6. Satb1
7. Sox17
8. Tpgs2
9. Hsf1
10. Pds5b

When assessing the overall sample contribution through their summary eigengenes it appears that for **yellow** all spleen had a positive effect, intestine had a negative effect, all independent of treatment, and skeletal muscle mostly had a negative effect with two 2DG samples having a positive effect. Intestine and muscle had mostly under-expressed genes while spleen had mostly over-expressed genes.
