==> Old_Rmd_files/08_C57BL6J_2DG_WGCNA_frequency_of_pathways_across_modules_related_to_traits_SM_SI_spleen_4wks.Rmd <==
---
title: "Frequency of Pathways by Module-Trait Relationship Skeletal Muscle, Small Intestine, and Spleen 4wks"
author: Ann Wells
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
 html_document:
   keep_md: TRUE
   code_folding: hide
   collapsed: no
   toc: yes
   toc_float: yes
   self_contained: no
   
---
<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.5/css/jquery.dataTables.min.css">
<script src="https://code.jquery.com/jquery-2.1.2.min.js"></script>
<script src="https://cdn.datatables.net/1.10.5/js/jquery.dataTables.min.js"></script>

<script type="text/javascript">
         $(document).ready(function() {
             $("table").DataTable();
         } );
</script>

# Introduction and Data files

```{r setup, echo=FALSE, warning=FALSE, message=FALSE}
options(stringsAsFactors = F)
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
knitr::opts_knit$set(progress=FALSE)
```

This dataset contains nine tissues(heart, hippocampus, hypothalamus, kidney, liver, prefrontal cortex, skeletal muscle, small intestine, and spleen) from C57BL/6J mice that were fed 2-deoxyglucose (6g/L) through their drinking water for 96hrs or 4wks. 96hr mice were given their 2DG treatment 2 weeks after the other cohort started the 4 week treatment. The organs from the mice were harvested and processed for proteomics, metabolomics, and transcriptomics. The data in this document pertains to the transcriptomics data only. The counts that were used were FPKM normalized before being log transformed. It was determined that sample A113 had low RNAseq quality and through further analyses with PCA, MA plots, and clustering was an outlier and will be removed for the rest of the analyses performed. This document will determine the contribution of each main effect or their combination to each module identified, as well as the relationship of each gene within the module and its significance.

```{r load packages}
needed.packages <- c("tidyverse", "here", "functional", "gplots", "dplyr", "GeneOverlap", "R.utils", "reshape2","magrittr","data.table", "RColorBrewer","preprocessCore", "ARTool","emmeans", "phia", "gProfileR", "WGCNA","plotly", "pheatmap","downloadthis")
for(i in 1:length(needed.packages)){library(needed.packages[i], character.only = TRUE)}

source(here("source_files","WGCNA_source.R"))
```

```{r}
tdata.FPKM.sample.info <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_phenotypes.RData"))

tdata.FPKM <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_numeric.RData"))

log.tdata.FPKM <- log(tdata.FPKM + 1)
log.tdata.FPKM <- as.data.frame(log.tdata.FPKM)

log.tdata.FPKM.sample.info <- cbind(log.tdata.FPKM, tdata.FPKM.sample.info[,27238:27240])

log.tdata.FPKM.sample.info <- log.tdata.FPKM.sample.info %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.subset <- log.tdata.FPKM[,colMeans(log.tdata.FPKM != 0) > 0.5] 

log.tdata.FPKM.subset <- log.tdata.FPKM.subset %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.sample.info.subset.SM.SI.spleen.4wks <- log.tdata.FPKM.sample.info %>% rownames_to_column() %>% filter(Tissue %in% c("Skeletal Muscle", "Small Intestine", "Spleen")) %>% filter(Time == "4 wks") %>% column_to_rownames()

log.tdata.FPKM.subset <- subset(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.4wks, select = -c(Time,Treatment,Tissue))
  
WGCNA.pathway <-readRDS(here("Data","SM.SI.spleen.4wks","Chang_B6_96hr_4wk_gprofiler_pathway_annotation_list_SM_SI_spleen_4wks_WGCNA.RData"))

Matched<-readRDS(here("Data","SM.SI.spleen.4wks","Annotated_genes_in_SM_SI_spleen_4wks_WGCNA_Chang_B6_96hr_4wk.RData"))
```

```{r}
module.names <- Matched$X..Module.
name <- strsplit(module.names,"_")
samples <-c()
for(i in 1:length(name)){
samples[[i]] <- name[[i]][2]
}  
name <- strsplit(samples,"\"")
name <- unlist(name)

Treatment <- unclass(as.factor(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.4wks[,27238]))
Tissue <- unclass(as.factor(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.4wks[,27239]))
Treat.Tissue <- paste0(Treatment,Tissue)
phenotype <- data.frame(cbind(Tissue, Treatment, Treat.Tissue))
nSamples <- nrow(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.4wks)

MEs0 <- read.csv(here("Data","SM.SI.spleen.4wks","log.tdata.FPKM.sample.info.subset.SM.SI.spleen.4wks.WGCNA.module.eigens.csv"),header = T, row.names = 1)
name <- strsplit(names(MEs0),"_")
samples <-c()
for(i in 1:length(name)){
samples[[i]] <- name[[i]][2]
}  
name <- strsplit(samples,"\"")
name <- unlist(name)
colnames(MEs0) <-name
MEs <- orderMEs(MEs0)
moduleTraitCor <- cor(MEs, phenotype, use = "p");
moduleTraitPvalue <- corPvalueStudent(moduleTraitCor, nSamples)
```

```{r}
  tissue.pathways <- list()
  treat.pathways <- list()
  tissue.treat.pathways <- list()
  
  for(i in 1:nrow(moduleTraitPvalue)){
    if(moduleTraitPvalue[i,1] < 0.05){
      tissue.pathways[[i]] <- WGCNA.pathway[[i]]
      }
    if(moduleTraitPvalue[i,2] < 0.05){
      treat.pathways[[i]] <- WGCNA.pathway[[i]]
      }
    if(moduleTraitPvalue[i,3] < 0.05){
      tissue.treat.pathways[[i]] <- WGCNA.pathway[[i]]  
      }
    }
```

## Pathway Frequency Across Modules Correlated to Each Trait {.tabset .tabset-fade .tabset-pills}

### Tissue
```{r, results='asis'}
count.pathways.traits(tissue.pathways)
```
