==> Old_Rmd_files/14_C57BL6J_2DG_genes_in_pathways_ANOVA_SM_SI_spleen_96hrs.Rmd <==
---
title: "Pathway Analysis SM.SI.spleen.96hrs, Small Intestine, and Spleen 96hrs"
author: Ann Wells
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output:
 html_document:
   keep_md: TRUE
   code_folding: hide
   collapsed: no
   toc: yes
   toc_float: yes
   self_contained: no
   
---

<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.5/css/jquery.dataTables.min.css">
<script src="https://code.jquery.com/jquery-2.1.2.min.js"></script>
<script src="https://cdn.datatables.net/1.10.5/js/jquery.dataTables.min.js"></script>

<script type="text/javascript">
         $(document).ready(function() {
             $("table").DataTable();
         } );
</script>

# Introduction

```{r setup, echo=FALSE, warning=FALSE, message=FALSE}
options(stringsAsFactors = F)
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
knitr::opts_knit$set(progress=FALSE)

```

This dataset contains nine tissues (heart, hippocampus, hypothalamus, kidney, liver, prefrontal cortex, skeletal muscle, small intestine, and spleen) from C57BL/6J mice that were fed 2-deoxyglucose (6g/L) through their drinking water for 96hrs or 4wks. 96hr mice were given their 2DG treatment 2 weeks after the other cohort started the 4 week treatment. The organs from the mice were harvested and processed for proteomics, metabolomics, and transcriptomics. The data in this document pertains to the transcriptomics data only. The counts that were used were FPKM normalized before being log transformed. It was determined that sample A113 had low RNAseq quality and through further analyses with PCA, MA plots, and clustering was an outlier and will be removed for the rest of the analyses performed. This document will determine which pathways are significantly altered by each module.

```{r load packages}
needed.packages <- c("tidyverse", "here", "functional", "gplots", "dplyr", "GeneOverlap", "R.utils", "reshape2","magrittr","data.table", "RColorBrewer","preprocessCore", "ARTool","emmeans", "phia", "gprofiler2","rlist", "plotly","kableExtra","foreach","doParallel","doFuture","downloadthis")
for(i in 1:length(needed.packages)){library(needed.packages[i], character.only = TRUE)}

source(here("source_files","WGCNA_source.R"))
```

```{r}
WGCNA.pathway <-readRDS(here("Data","SM.SI.spleen.96hrs","Chang_B6_96hr_4wk_gprofiler_pathway_annotation_list_SM_SI_spleen_96hrs_WGCNA.RData"))

WGCNA.gene <-readRDS(here("Data","SM.SI.spleen.96hrs","Chang_B6_96hr_4wk_gprofiler_gene_annotation_list_SM_SI_spleen_96hrs_WGCNA.RData"))

tdata.FPKM.sample.info <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_phenotypes.RData"))

tdata.FPKM <- readRDS(here("Data","20190406_RNAseq_B6_4wk_2DG_counts_numeric.RData"))

log.tdata.FPKM <- log(tdata.FPKM + 1)
log.tdata.FPKM <- as.data.frame(log.tdata.FPKM)

log.tdata.FPKM.sample.info <- cbind(log.tdata.FPKM, tdata.FPKM.sample.info[,27238:27240])  

log.tdata.FPKM.sample.info <- log.tdata.FPKM.sample.info %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.subset <- log.tdata.FPKM[,colMeans(log.tdata.FPKM != 0) > 0.5] 

log.tdata.FPKM.sample.info.subset <- cbind(log.tdata.FPKM.subset,tdata.FPKM.sample.info[,27238:27240])  

log.tdata.FPKM.sample.info <- log.tdata.FPKM.sample.info %>% rownames_to_column() %>% filter(rowname != "A113") %>% column_to_rownames()

log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs <- log.tdata.FPKM.sample.info.subset %>% rownames_to_column() %>% filter(Tissue %in% c("Skeletal Muscle", "Small Intestine", "Spleen")) %>% filter(Time == "96 hrs") %>% column_to_rownames()

ensembl.location <- readRDS(here("Data","Ensembl_gene_id_and_location.RData"))

Matched<-readRDS(here("Data","SM.SI.spleen.96hrs","Annotated_genes_in_SM_SI_spleen_96hrs_WGCNA_Chang_B6_96hr_4wk.RData"))


module.names <- Matched$X..Module.
name <- strsplit(module.names,"_")
samples <-c()
for(i in 1:length(name)){
samples[[i]] <- name[[i]][2]
}  
name2 <- stringr::str_split(samples,"\"")
mod.names <- c()
for(i in 1:length(name2)){
   mod.names[[i]] <- name2[[i]][1]
}
name2 <- unlist(mod.names)
modules <- unique(mod.names)
modules <- unlist(modules)
```

## ANOVA {.tabset .tabset-fade .tabset-pills}
```{r}
log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs$Time <- as.factor(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs$Time)
log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs$Treatment <- as.factor(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs$Treatment)
log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs$Tissue <- as.factor(log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs$Tissue)

genes <- unlist(WGCNA.gene)
genes <- genes[-c(9870,25160)]
```

```{r, results='asis', cache=TRUE, eval=FALSE}

cl <- makeCluster(30)
registerDoParallel(cl)

stat.table <- foreach(i = 1:length(genes),.packages = c("ARTool","tidyverse","kableExtra")) %dopar% {
   m2 <- art(data = log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs, log.tdata.FPKM.sample.info.subset.SM.SI.spleen.96hrs[,genes[i]] ~ Tissue*Treatment)

   model <- anova(m2)

   adjust <- p.adjust(model$`Pr(>F)`, method = "BH")

   return(cbind(model,adjust))
}

stopCluster(cl)

saveRDS(stat.table,here("Data","SM.SI.spleen.96hrs","Chang_B6_96hr_4wk_WGCNA_gene_ANOVA_ART_SM_SI_spleen_96hrs.RData"))
```

```{r, results='asis', cache = TRUE}

stat.table <- readRDS(here("Data","SM.SI.spleen.96hrs","Chang_B6_96hr_4wk_WGCNA_gene_ANOVA_ART_SM_SI_spleen_96hrs.RData"))

external.gene.name <- ensembl.location$external_gene_name[match(genes,ensembl.location$ensembl_gene_id)] 
names(stat.table) <- external.gene.name

system.time(Pathway.gene.ANOVA(modules, WGCNA.gene))
```
